+++ src/collective/contentsections/sections/cards/content.py
@@ -85,7 +85,7 @@ class ICardsSection(ISection):
     )
 
     directives.widget("cards", DataGridFieldFactory, allow_reorder=True)
-    directives.order_before(group_size="is_full_width")
+    directives.order_before(group_size="container_width")
 
 
 @implementer(ICardsSection)